Use "Simulation" to run the board including the LCD screens (#1)
* Create Simulate button
Simulate brings up screen.  Button areas on screen are active (1,2,3,4).
Use commands "button(x)" and "getscreen" active on board (need version 2.2.3)
* Simulate State machine with pictures
Import all pictures from the India Demo, and run through them with the "Simulation" screen.
Note that the state machine is on the board, not in this Windows app.  The Simulation just sends button pushes to the board, and the board then updates the "current screen" to display.