v1.1.34 - Handle O2 sensor status other than 0 and 1
O2 sensor status "2" (50%) means the sensor heater has a problem.  It could be the interface board isn't generating the heater voltage, or it could be a short of the sensor heater coils.
O2 sensor status "3" (75%) means the sensor doesn't match the interface board, or the sensor is damaged through misuse, or the O2 is below 1%.  Show these "red" on the screen.